+++ visualization/app/codeCharta/state/injector.service.ts
@@ -1,5 +1,7 @@
 // Plop: Append service import here
 import { FilesService } from "./store/fileSettings/files/files.service"
+import { IsLoadingFileService } from "./store/appSettings/isLoadingFile/isLoadingFile.service"
+import { IsLoadingMapService } from "./store/appSettings/isLoadingMap/isLoadingMap.service"
 import { MapSizeService } from "./store/treeMap/mapSize/mapSize.service"
 import { MapColorsService } from "./store/appSettings/mapColors/mapColors.service"
 import { ResetCameraIfNewFileIsLoadedService } from "./store/appSettings/resetCameraIfNewFileIsLoaded/resetCameraIfNewFileIsLoaded.service"
@@ -39,6 +41,8 @@ export class InjectorService {
 		// tslint:disable:no-unused-variable
 		// We have to inject the services somewhere
 		// Plop: Append service injection here
+		private isLoadingFileService: IsLoadingFileService,
+		private isLoadingMapService: IsLoadingMapService,
 		private filesService: FilesService,
 		private mapSizeService: MapSizeService,
 		private mapColorsService: MapColorsService,