+++ visualization/app/codeCharta/state/state.module.ts
@@ -1,5 +1,7 @@
 import angular from "angular"
 // Plop: Append module import here
+import { IsLoadingFileService } from "./store/appSettings/isLoadingFile/isLoadingFile.service"
+import { IsLoadingMapService } from "./store/appSettings/isLoadingMap/isLoadingMap.service"
 import { FilesService } from "./store/fileSettings/files/files.service"
 import { MapSizeService } from "./store/treeMap/mapSize/mapSize.service"
 import { MapColorsService } from "./store/appSettings/mapColors/mapColors.service"
@@ -38,7 +40,6 @@ import { SearchPanelService } from "./searchPanel.service"
 import { NodeSearchService } from "./nodeSearch.service"
 import { FileStateService } from "./fileState.service"
 import { MetricService } from "./metric.service"
-import { LoadingStatusService } from "./loadingStatus.service"
 import "../codeCharta.module"
 import _ from "lodash"
 import { IsPresentationModeService } from "./store/appSettings/isPresentationMode/isPresentationMode.service"
@@ -46,6 +47,8 @@ import { IsPresentationModeService } from "./store/appSettings/isPresentationMod
 angular
 	.module("app.codeCharta.state", ["app.codeCharta"])
 	// Plop: Append service name here
+	.service(_.camelCase(IsLoadingFileService.name), IsLoadingFileService)
+	.service(_.camelCase(IsLoadingMapService.name), IsLoadingMapService)
 	.service(_.camelCase(FilesService.name), FilesService)
 	.service(_.camelCase(MapSizeService.name), MapSizeService)
 	.service(_.camelCase(MapColorsService.name), MapColorsService)
@@ -85,4 +88,3 @@ angular
 	.service(_.camelCase(NodeSearchService.name), NodeSearchService)
 	.service(_.camelCase(FileStateService.name), FileStateService)
 	.service(_.camelCase(MetricService.name), MetricService)
-	.service(_.camelCase(LoadingStatusService.name), LoadingStatusService)