+++ visualization/app/codeCharta/state/store/fileSettings/markedPackages/markedPackages.service.ts
@@ -12,7 +12,7 @@ export class MarkedPackagesService implements StoreSubscriber {
 	private static MARKED_PACKAGES_CHANGED_EVENT = "marked-packages-changed"
 
 	constructor(private $rootScope: IRootScopeService, private storeService: StoreService) {
-		StoreService.subscribe($rootScope, this)
+		StoreService.subscribe(this.$rootScope, this)
 	}
 
 	public onStoreChanged(actionType: string) {