+++ CHANGELOG.md
@@ -9,11 +9,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Added 🚀
 
+- RawTextParser for analysis #660
+- IndentationLevel as metric for RawTextParser #660
+- Show additional Pairing Rate of Selected Building, simultaneously to the currently hovered Buildings #736
+
 ### Changed
 
 ### Removed 🗑
 
 ### Fixed 🐞
+
+- Path prefix handling in tokeiimporter #841
 
 ### Chore 👨‍💻 👩‍💻